+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/PrometheusExporterMiddleware.cs
@@ -18,7 +18,7 @@ namespace OpenTelemetry.Exporter;
 /// </summary>
 internal sealed class PrometheusExporterMiddleware
 {
-    private readonly PrometheusExporter exporter;
+    private readonly PrometheusExporter? exporter;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="PrometheusExporterMiddleware"/> class.
@@ -32,7 +32,10 @@ internal sealed class PrometheusExporterMiddleware
 
         if (!meterProvider.TryFindExporter(out PrometheusExporter? exporter))
         {
-            throw new ArgumentException("A PrometheusExporter could not be found configured on the provided MeterProvider.");
+            // If the SDK is disabled, just configure a no-op exporter
+            exporter = meterProvider is OpenTelemetrySdk.NoopMeterProvider
+                ? null
+                : throw new ArgumentException("A PrometheusExporter could not be found configured on the provided MeterProvider.");
         }
 
         this.exporter = exporter;
@@ -53,6 +56,14 @@ internal sealed class PrometheusExporterMiddleware
 
         try
         {
+            if (this.exporter is null)
+            {
+                // The SDK was disabled, so we don't have an exporter to use.
+                // Just return 200 OK with no content as an effective no-op.
+                response.StatusCode = StatusCodes.Status200OK;
+                return;
+            }
+
             using var requestCancelled = new CancellationTokenSource();
 
             int? scrapeTimeoutSeconds = null;